Add files via upload
v1.1 update log

New Features

Calendar Navigation - A calendar visual has been added as well as a clearer view of the current date that is reflected in the chart. A "Today" button has been added to jump to the current day.

Weekly View - The user can toggle between a daily and a weekly view of the event reports. In the weekly view, the timeline changes the x-axis to a Sun-Sat timeline and the y-axis lists the companies occupying rooms on certain days.

Search Bar - A search bar has been added to filter the events to a specific company or event name within the chart. Type in the name of the company or event and it will auto-suggest what you are looking for. 

Bug Fixes

- Updated What's Happening tab to include all events happening within the day. Previously it listed only immediate upcoming events within a few hours and not the entire day's schedule.